# backend/docs/swagger.yaml
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      messages:
        items: {}
        type: array
      meta: {}
    type: object
host: localhost:8080
info:
  contact:
    email: krahmadani1@gmail.com
    name: Kukuh Rahmadani
  description: This is a sample API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aplikasi Music API
  version: "1.0"
paths:
  /api/v1/auth/me:
    get:
      description: API for get me
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Me
      tags:
      - Authentication
  /api/v1/login:
    post:
      description: API for do login
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Do login
      tags:
      - Authentication
  /api/v1/register:
    post:
      description: API for register
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register
      tags:
      - Authentication
  /music:
    get:
      consumes:
      - application/json
      description: Get list of tracks with search and pagination
      parameters:
      - description: Search by title or artist
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get paginated tracks
      tags:
      - Music
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer {TOKEN}'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"